+++ Cargo.toml
@@ -86,13 +86,13 @@ async-stream = "0.3.5"
 bytes = "1.4.0"
 chrono = "0.4.26"
 chrono-tz = "0.8.3"
+comfy-table = "7.1.0"
 futures = "0.3.28"
 html-escape = "0.2.13"
 indexmap = "2.0.0"
 itertools = "0.11"
 num-derive = "0.3.3"
 num-traits = "0.2"
-prettytable-rs = "0.10"
 rand = "^0.8"
 rayon = "1.7.0"
 rstest = "0.18.2"
@@ -102,7 +102,6 @@ tokio = {version = "1.32.0", features = ["net", "time", "bytes", "process", "sig
 tokio-stream = {version = "0.1.14", features = ["fs"]}
 tokio-util = "0.7.8"
 url = "2.4.0"
-comfy-table = "7.1.0"
 
 [workspace.dependencies.arrow2]
 # branch = "jay/fix-parquet-timezone-parsing"
