+++ src/daft-core/Cargo.toml
@@ -4,6 +4,7 @@ base64 = "0.21.5"
 bincode = {workspace = true}
 chrono = {workspace = true}
 chrono-tz = {workspace = true}
+comfy-table = {workspace = true}
 common-error = {path = "../common/error", default-features = false}
 dyn-clone = "1.0.16"
 fnv = "1.0.7"
@@ -14,7 +15,6 @@ log = {workspace = true}
 ndarray = "0.15.6"
 num-derive = {workspace = true}
 num-traits = {workspace = true}
-prettytable-rs = {workspace = true}
 pyo3 = {workspace = true, optional = true}
 pyo3-log = {workspace = true}
 rand = {workspace = true}
