+++ src/daft-core/src/series/mod.rs
@@ -4,6 +4,7 @@ mod ops;
 mod serdes;
 mod series_like;
 use std::{
+    borrow::Cow,
     fmt::{Display, Formatter, Result},
     sync::Arc,
 };
@@ -11,6 +12,7 @@ use std::{
 use crate::{
     array::ops::{from_arrow::FromArrow, full::FullNull},
     datatypes::{DataType, Field},
+    utils::display_table::make_comfy_table,
     with_match_daft_types,
 };
 use common_error::DaftResult;
@@ -77,47 +79,19 @@ impl Series {
         }
     }
 
-    pub fn to_prettytable(&self) -> prettytable::Table {
-        let mut table = prettytable::Table::new();
-
-        let header =
-            prettytable::Cell::new(format!("{}\n{}", self.name(), self.data_type()).as_str())
-                .with_style(prettytable::Attr::Bold);
-        table.add_row(prettytable::Row::new(vec![header]));
-
-        let head_rows;
-        let tail_rows;
-
-        if self.len() > 10 {
-            head_rows = 5;
-            tail_rows = 5;
-        } else {
-            head_rows = self.len();
-            tail_rows = 0;
-        }
-
-        for i in 0..head_rows {
-            let row = vec![self.str_value(i).unwrap()];
-            table.add_row(row.into());
-        }
-        if tail_rows != 0 {
-            let row = vec!["..."];
-            table.add_row(row.into());
-        }
-
-        for i in 0..tail_rows {
-            let row = vec![self.str_value(self.len() - tail_rows - 1 + i).unwrap()];
-            table.add_row(row.into());
-        }
-
-        table
+    pub fn to_comfy_table(&self) -> comfy_table::Table {
+        make_comfy_table(
+            vec![Cow::Borrowed(self.field())].as_slice(),
+            Some([self].as_slice()),
+            Some(32),
+        )
     }
 }
 
 impl Display for Series {
     // `f` is a buffer, and this method must write the formatted string into it
     fn fmt(&self, f: &mut Formatter) -> Result {
-        let table = self.to_prettytable();
+        let table = self.to_comfy_table();
         write!(f, "{table}")
     }
 }
