+++ src/daft-table/Cargo.toml
@@ -1,13 +1,11 @@
 [dependencies]
 arrow2 = {workspace = true}
+comfy-table = {workspace = true}
 common-error = {path = "../common/error", default-features = false}
 daft-core = {path = "../daft-core", default-features = false}
 daft-dsl = {path = "../daft-dsl", default-features = false}
 html-escape = {workspace = true}
 num-traits = {workspace = true}
-prettytable-rs = {workspace = true}
-comfy-table = {workspace = true}
-
 pyo3 = {workspace = true, optional = true}
 pyo3-log = {workspace = true}
 rand = {workspace = true}
